Fix few CSS styles
